Make ground check work on different angled ground and minor fixes on turning
Make ground check work on different angled ground by adding a child
object which has a box collider that the car is on top of. New box
collider has player's old drag values which are changed in PlayerInput
script. Player's turning speed on low velocity is also now in relation
to the velocity itself.